+++ source/client/cin.h
@@ -23,14 +23,14 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 void CIN_LoadLibrary( bool verbose );
 void CIN_UnloadLibrary( bool verbose );
 
-struct cinematics_s *CIN_Open( const char *name, unsigned int start_time,
+struct cinematics_s *CIN_Open( const char *name, int64_t start_time,
 							   int flags, bool *yuv, float *framerate );
 
 bool CIN_HasOggAudio( struct cinematics_s *cin );
 
 const char *CIN_FileName( struct cinematics_s *cin );
 
-bool CIN_NeedNextFrame( struct cinematics_s *cin, unsigned int curtime );
+bool CIN_NeedNextFrame( struct cinematics_s *cin, int64_t curtime );
 
 uint8_t *CIN_ReadNextFrame( struct cinematics_s *cin, int *width, int *height,
 							int *aspect_numerator, int *aspect_denominator, bool *redraw );
@@ -41,6 +41,6 @@ ref_yuv_t *CIN_ReadNextFrameYUV( struct cinematics_s *cin, int *width, int *heig
 bool CIN_AddRawSamplesListener( struct cinematics_s *cin, void *listener,
 								cin_raw_samples_cb_t rs, cin_get_raw_samples_cb_t grs );
 
-void CIN_Reset( struct cinematics_s *cin, unsigned int cur_time );
+void CIN_Reset( struct cinematics_s *cin, int64_t cur_time );
 
 void CIN_Close( struct cinematics_s *cin );
